shim: Factorize shim execution code
Shim implementations can share the shim binary execution code.
What differs between them is the arguments list building routine.
