+++ kata_shim.go
@@ -18,9 +18,6 @@ package virtcontainers
 
 import (
 	"fmt"
-	"os"
-	"os/exec"
-	"syscall"
 )
 
 type kataShim struct{}
@@ -66,45 +63,5 @@ func (s *kataShim) start(pod Pod, params ShimParams) (int, error) {
 		args = append(args, "-d")
 	}
 
-	cmd := exec.Command(args[0], args[1:]...)
-	cmd.Env = os.Environ()
-
-	if !params.Detach {
-		cmd.Stdin = os.Stdin
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-	}
-
-	var f *os.File
-	var err error
-	if params.Console != "" {
-		f, err = os.OpenFile(params.Console, os.O_RDWR, consoleFileMode)
-		if err != nil {
-			return -1, err
-		}
-
-		cmd.Stdin = f
-		cmd.Stdout = f
-		cmd.Stderr = f
-		cmd.SysProcAttr = &syscall.SysProcAttr{
-			// Create Session
-			Setsid: true,
-
-			// Set Controlling terminal to Ctty
-			Setctty: true,
-			Ctty:    int(f.Fd()),
-		}
-
-	}
-	defer func() {
-		if f != nil {
-			f.Close()
-		}
-	}()
-
-	if err := cmd.Start(); err != nil {
-		return -1, err
-	}
-
-	return cmd.Process.Pid, nil
+	return startShim(args, params)
 }
